[test_nn] split multihead_attention from test_nn
